Authorization using gates

+++ app/Http/Controllers/QuestionController.php
@@ -8,6 +8,10 @@ use App\Question;
 
 class QuestionController extends Controller
 {
+
+        public function __construct() {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -62,7 +66,13 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-       return view('questions.edit',compact('question'));
+        if(\Gate::denies('update-question',$question)){
+            
+            abort(403,"access denied");
+        };
+        return view('questions.edit',compact('question'));
+        
+       
     }
 
     /**
@@ -87,6 +97,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question',$question)){
+            
+            abort(403,"access denied");
+        };
         $question->delete();
         return redirect()->route('questions.index')->with('success',"Your question has been deleted");
     }
